rebased @ashi009's changes with refactored master.
